plugin: sort tests by topology name
This is needed in order to allow one-time topology setup and teardown.

If the tests would not be sorted by topology, we would have to run
topology setup and teardown for each test, which has its use cases
but in most majority of cases you want to setup environment for
given topology only once and then run all tests for this topology
in order to speed things up.

We will provide hooks to do both - one-time setup and per-test setup.