multihost: add mh prefix to selected fields
Domain and config are not likely to be used by users but on the
other hand these name are very likely to be used differently such
as "custom domain config" or "dns domain" or "identity domain"
and so on. Therefore it makes sense to prefix them.

Other fields are expected to be used so it does not makes sense
to prefix them but rather keep them short and clear.