mh: support extra artifacts in host and utility as well
It is possible to set additional artifacts to collect from
MultihostUtility and MultihostHost objects as well.

It is also possible to make the list more dynamic using the
prepare_artifacts method that can be use to produce more
artifacts or check which extra artifacts were produced
(like selinux errors or coredumps).