multihost: setup and teardown MultihostUtility using a dependency tree
So far, the utilities where setup and teardown in a random order (in
reality is was in the order of their presence in the code, but this
is python's implementation detail).

Since utilities may require other utilities, it is fundamental that
we call the methods in a correct order. We now sort the list based
on a computed dependency tree.

- We sort the list by class name
- Then add utilities that have zero dependencies
- Then add utilities which dependency is already satisfied in a loop