multihost: simplify setup_when_used
The main purpose of setup_when_used and teardown_when_used is to
postpone utility setup to a point when it is actually used in order
to avoid costly operations on stuff the is rarely used in tests.

However, the implementation of additional methods adds lots of
unnecessary complexity. With this patch, postponed setup is kept
but the *_when_used methods were removed, setup() and teardown()
are called instead.

There is a class decorator To advertise the desire to postpone setup,
for example:

```
@mh_utility_postpone_setup
class MyUtility(MultihostUtility):
    pass
```

Note that LinuxTrafficControl and Firewall classes are no longer using
postponed setup intentionally. We should not enforce it, the user of
this class can decide on its own if the setup should be postponed or not.