multihost: remove ssh from MultihostRole
We will add a generic connection API so other methods besides ssh
can be used. This call is used to login to the host as given user
and to test the ssh connection. It is test specific and if it is
needed it should be provided by the test framework built on
pytest-mh, not by pytest-mh itself.