Increase minimum LibsshSession timeout
When the timeout is set to 1 in LibsshSession(timeout=1), pylibssh
no longer returns an exception for a wrong password SSH connect()

    with pytest.raises(SSHAuthenticationError):
        client.ssh("user1", "Wrong").connect()

Increase the minimum timeout a couple seconds here.

Regarding 'LibsshSession(timeout=1)', a very low value must be set
here, to allow the timeout set in

        client.host.conn.run(..., timeout=X)

to work as expected. This is due to code becoming blocked in libssh
which is C not Python, but the Python signal is delayed until we get
back to the Python code