markers: new preferred_topology marker
Reducing test running time and system resources, we can now mark
a test with @pytest.mark.preferred_topology(KnownTopology.topology),
making the topology the single and default run unless the marker is
overridden. This is intended to be used for tests that do not need
to cover all topologies all the time.